(#40) separated classes in SchurInfo into seperate header files
